requirements.txt
aiohttp==2.2.3
async-timeout==1.2.1
chardet==3.0.4
Django==1.11.3
djangorestframework==3.6.3
multidict==3.1.0
pytz==2017.2
telepot==12.1
urllib3==1.21.1
yarl==0.11.0
django-environ==0.4.3
celery==4.1.0